added print, convert dataset in utils and jsonl files
