+++ legal_ner/utils/dataset.py
@@ -3,7 +3,7 @@ import json
 from torch.utils.data import Dataset
 import numpy as np
 from transformers import AutoTokenizer, RobertaTokenizerFast
-
+from datasets import load_dataset, Dataset, DatasetDict
 from utils.utils import match_labels
 
 import spacy
@@ -85,3 +85,15 @@ class LegalNERTokenDataset(Dataset):
                 inputs["labels"] = labels[: inputs["attention_mask"].shape[0]]
 
         return inputs
+    
+
+## FOR SPAN
+def load_legal_ner(data_folder):
+    ret = {}
+    for split_name in ['TRAIN', 'DEV']:
+        data = []
+        with open(f"{data_folder}/NER_{split_name}/NER_{split_name}_ALL_OT.jsonl", 'r') as reader:
+            for line in reader:
+                data.append(json.loads(line))
+        ret[split_name.lower()] = Dataset.from_list(data)
+    return DatasetDict(ret)
